requirements_simple.txt
# Core Flask Dependencies
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.2
Flask-Mail==0.9.1
Flask-Bcrypt==1.0.1

# Database
SQLAlchemy==2.0.43

# PDF Generation
ReportLab==4.2.2

# QR Code Generation
qrcode==7.4.2
Pillow==10.4.0

# Payment Processing
stripe==10.12.0

# Data Processing (Excel support without pandas)
openpyxl==3.1.5

# Security & Rate Limiting
Flask-Limiter==3.8.0

# Environment Management
python-dotenv==1.1.1

# Utilities
Werkzeug==3.1.3
WTForms==3.2.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
click==8.2.1
blinker==1.9.0
colorama==0.4.6
